CLI args, dump, code generation
The code generation has been improved and bugs fixed, anything which can be
named currently does not work (constants, variables and functions) as the
search and scoping mechanisms need to be added.

The program now accepts a limited set of arguments, instead of just reading
from stdin.

A dump function has been added to the generated code can be viewed.