Stage 1 Complete :- Test Program Compiles
It appears the first test program works, "ex1.pas", by inspection of the
generated code. There is still a lot of work to be done on this compiler, even
if it is only a toy. For example, all variables are stored in a global variable
list, and the virtual machine has not been implemented as of yet.

In this commit:

* bugs in code generation have been fixed (UNARY_TERM did not generate code
for EXPRESSIONS).
* variables can be looked up, used, and allocated
* very limited sanity checks have been added (redeclaring a variable in the
same scope should be checked for, which is not done at the moment)
* more information is now stored in the token node (line number, what type of
identifier it is, if it is one)
* dump now prints out variables used
* code_t object is now freed after use