Code refactoring, printing out symbols
The code has been refactored, memory allocations are a little bit more
efficient, and exit is called less, but needs removing more.

Code generation has been improved for procedure entry.

A list of symbols with their scopes can now be exported, which may be used in
the future for linking.